requirements.txt
flet==0.26.0
flet-web==0.26.0
dotenv
psycopg2-binary